requirements.txt
instagrapi==2.0.0
python-dotenv==1.0.0
rich==13.7.0
cryptography==41.0.7